backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.3
python-dotenv==1.0.0
yt-dlp==2024.1.12
requests==2.31.0
python-multipart==0.0.6
